Extract HeaderIncrement
